refactor: 디테일 페이지 react-query에서 prefetching할 수 있도록 변경
