+++ app/(home)/(board)/[menu]/[id]/layout.tsx
@@ -1,10 +1,12 @@
-import { Fragment, ReactNode } from 'react';
+import { ReactNode } from 'react';
 import type { Metadata } from 'next';
+import { notFound } from 'next/navigation';
 
 import { BASE_API_URL, BASE_FRONT_URL } from '@/constant/api';
 import { allRoutes } from '@/constant/Routes';
 import { PageParams } from '@/constant/type';
 import { PostData } from '@/apis/posts';
+import ContainerBox from '@/components/common/ContainerBox';
 
 export const generateMetadata = async ({
   params,
@@ -12,20 +14,26 @@ export const generateMetadata = async ({
   params: PageParams;
 }): Promise<Metadata> => {
   const { id, menu } = await params;
+
   const url = `${BASE_API_URL}/posts/${id}`;
-  const { post } = (await fetch(url).then((res) => res.json()))
-    .data as PostData;
+  const { data } = (await fetch(url).then((res) => res.json())) as {
+    data: PostData;
+  };
+
+  if (!id || !menu || !data) {
+    notFound();
+  }
 
   const menuName = allRoutes.filter(({ link }) => link.includes(menu))[0];
 
   return {
-    title: `옥동 || ${post.title}`,
+    title: `옥동 || ${data.post.title}`,
     description: `한국춘란 산채품 전문 직거래장터 - ${menuName.name}`,
-    metadataBase: new URL(`${BASE_FRONT_URL}/${post.marketType}`),
+    metadataBase: new URL(`${BASE_FRONT_URL}/${data.post.marketType}`),
     openGraph: {
-      title: `옥동 || ${post.title}`,
-      description: post.contents,
-      ...(post.image[0] && { images: [post.image[0]] }),
+      title: `옥동 || ${data.post.title}`,
+      description: data.post.contents,
+      ...(data.post.image[0] && { images: [data.post.image[0]] }),
     },
     keywords: [
       '옥동',
@@ -35,21 +43,15 @@ export const generateMetadata = async ({
       '직거래',
       '산채',
       '약초',
-      post.title,
-      post.variant,
-      post.place || '춘란',
+      data.post.title,
+      data.post.variant,
+      data.post.place || '춘란',
     ],
   };
 };
 
-const DetailLayout = async ({
-  children,
-  params,
-}: {
-  children: ReactNode;
-  params: { menu: string; id: string };
-}) => {
-  return <Fragment>{children}</Fragment>;
+const DetailLayout = async ({ children }: { children: ReactNode }) => {
+  return <ContainerBox>{children}</ContainerBox>;
 };
 
 export default DetailLayout;
