+++ app/(home)/(board)/[menu]/[id]/page.tsx
@@ -1,29 +1,17 @@
 'use client';
-import React, { Fragment, Suspense } from 'react';
-import { useQuery } from '@tanstack/react-query';
+import React, { Suspense } from 'react';
 
-import ContainerBox from '@/components/common/ContainerBox';
 import PostDetail from '@/components/Posts/PostDetail';
-// import { usePost } from '@/hooks/queries/usePosts';
 import { PageParams } from '@/constant/type';
-import { postQueryOptions } from '@/constant/queryOptions/postQueryOptions';
-import Loading from '@/components/common/Loading';
+import LoadingPage from '@/(home)/loading';
 
 const DetailPage = ({ params }: { params: PageParams }) => {
   const { id } = React.use(params);
 
   return (
-    <Fragment>
-      <Suspense
-        fallback={
-          <ContainerBox className='py-20'>
-            <Loading />
-          </ContainerBox>
-        }
-      >
-        <PostDetail postId={Number(id)} />
-      </Suspense>
-    </Fragment>
+    <Suspense fallback={<LoadingPage />}>
+      <PostDetail postId={Number(id)} />
+    </Suspense>
   );
 };
