+++ app/(home)/(board)/community/[menu]/[id]/layout.tsx
@@ -1,10 +1,12 @@
 import { Fragment, ReactNode } from 'react';
 import type { Metadata } from 'next';
+import { notFound } from 'next/navigation';
 
 import { PostData } from '@/apis/posts';
 import { BASE_API_URL, BASE_FRONT_URL } from '@/constant/api';
 import { allRoutes } from '@/constant/Routes';
 import { PageParams } from '@/constant/type';
+import ContainerBox from '@/components/common/ContainerBox';
 
 export const generateMetadata = async ({
   params,
@@ -13,19 +15,24 @@ export const generateMetadata = async ({
 }): Promise<Metadata> => {
   const { id, menu } = await params;
   const url = `${BASE_API_URL}/posts/${id}`;
-  const { post } = (await fetch(url).then((res) => res.json()))
-    .data as PostData;
+  const { data } = (await fetch(url).then((res) => res.json())) as {
+    data: PostData;
+  };
+
+  if (!id || !menu || !data) {
+    notFound();
+  }
 
   const menuName = allRoutes.filter(({ link }) => link.includes(menu))[0];
 
   return {
-    title: `옥동 || ${post.title}`,
+    title: `옥동 || ${data.post.title}`,
     description: `한국춘란 산채품 전문 직거래장터 - ${menuName.name}`,
-    metadataBase: new URL(`${BASE_FRONT_URL}/${post.marketType}`),
+    metadataBase: new URL(`${BASE_FRONT_URL}/${data.post.marketType}`),
     openGraph: {
-      title: `옥동 || ${post.title}`,
-      description: post.contents,
-      ...(post.image[0] && { images: [post.image[0]] }),
+      title: `옥동 || ${data.post.title}`,
+      description: data.post.contents,
+      ...(data.post.image[0] && { images: [data.post.image[0]] }),
     },
     keywords: [
       '옥동',
@@ -34,7 +41,7 @@ export const generateMetadata = async ({
       '산채품',
       '직거래',
       '산채',
-      post.title,
+      data.post.title,
       '산채기',
       '난자랑',
     ],
@@ -42,7 +49,7 @@ export const generateMetadata = async ({
 };
 
 const CommunityDetailLayout = ({ children }: { children: ReactNode }) => {
-  return <Fragment>{children}</Fragment>;
+  return <ContainerBox>{children}</ContainerBox>;
 };
 
 export default CommunityDetailLayout;
