+++ app/components/Community/CommDetail.tsx
@@ -1,137 +1,39 @@
 'use client';
-import { useEffect, useState } from 'react';
-import { usePathname, useRouter, useSearchParams } from 'next/navigation';
-import { IoArrowBack } from 'react-icons/io5';
+import { Fragment } from 'react';
+import { usePathname } from 'next/navigation';
+import { useQuery } from '@tanstack/react-query';
 
-import DateFormat from '@/utils/DateFormat';
-import ContainerBox from '@/components/common/ContainerBox';
-import PrevNextPost, {
-  PrevNextPostData,
-} from '@/components/Posts/PrevNextPost';
-import AutoHeightImageWrapper from '../common/Wrapper/AutoHeightImageWrapper';
-import HasLikes from '../Bookmark/BookmarkButton';
+import PrevNextPost from '@/components/Posts/PrevNextPost';
 import Comments from '../Comment/Comments';
 import EditorReadOnly from '../Editor/ReadOnly';
-import { deletePost } from '@/apis/posts';
-import { getUser } from '@/apis/users';
-import { Post } from '../common/Board/types';
-import { useAuthStore, User } from '@/stores/useAuthStore';
+import { DetailHeader } from '../Posts/PostDetail/DetailHeader';
+import { postQueryOptions } from '@/constant/queryOptions/postQueryOptions';
+import DetailSkeleton from '../Posts/PostDetail/DetailSkeleton';
+import { DetailImage } from '../Posts/PostDetail/DetailImage';
 
 interface CommDetailPageProps {
   postId: number;
-  data: {
-    previous: PrevNextPostData | null;
-    next: PrevNextPostData | null;
-    post: Post;
-  };
 }
 
-const CommDetailPage = ({ postId, data }: CommDetailPageProps) => {
-  const router = useRouter();
-  const { user } = useAuthStore();
-  const [userInfo, setUserInfo] = useState<User | null>(null);
-  const searchParams = useSearchParams();
-  const page = searchParams.get('page');
+const CommDetailPage = ({ postId }: CommDetailPageProps) => {
+  const { data, isLoading } = useQuery(postQueryOptions.getPost(postId));
 
   const path = usePathname().split('/');
   const pathname = path[3] ? path[2] : path[1];
 
-  useEffect(() => {
-    if (user) {
-      (async () => {
-        const userInfo = await getUser();
-        setUserInfo(userInfo ?? null);
-      })();
-    }
-  }, [user]);
+  if (isLoading) return <DetailSkeleton />;
 
-  const handleDeletePost = async (id: number) => {
-    const ok = window.confirm('이 게시물을 삭제하시겠습니까?');
-
-    if (!data) return;
-
-    if (ok) {
-      const res = await deletePost(id);
-      if (res.result === 'SUCCESS') {
-        router.push(`/${pathname}`);
-      }
-    }
-  };
-
-  const handleBack = () => {
-    const isCommunity = path[3] ? `/community/${pathname}` : `/${pathname}`;
-    const pageParam = page ? `?page=${page}` : '';
-    router.push(isCommunity + pageParam);
-  };
+  if (!data)
+    return (
+      <p className='py-20'>삭제된 게시물이거나 찾을 수 없는 게시물입니다.</p>
+    );
 
   return (
-    <ContainerBox>
-      {/* title */}
-      <div
-        className='border-b border-grayColor-400
-          flex gap-2 md:gap-4 justify-between items-center
-          py-3'
-      >
-        <div className='p-2 cursor-pointer' onClick={handleBack}>
-          <IoArrowBack size={18} />
-        </div>
-
-        <h2 className='text-lg font-bold flex-grow min-w-[120px] truncate'>
-          {data.post.title}
-        </h2>
-
-        {(userInfo?.id === data.post.userId || userInfo?.isAdmin) && (
-          <ul className='flex gap-2 text-gray-500 text-sm [&_li]:cursor-pointer'>
-            <li
-              onClick={() =>
-                router.push(`/community/${pathname}/edit/${postId}`)
-              }
-            >
-              편집
-            </li>
-            <li onClick={() => handleDeletePost(postId)}>삭제</li>
-          </ul>
-        )}
-      </div>
-
-      {/* post info */}
-      <ul className='flex gap-4 pt-2 pb-6 justify-end text-sm text-gray-500'>
-        <li className='flex flex-wrap gap-1 md:gap-2'>
-          <span className='flex-grow lg:flex-grow-0 font-semibold'>작성자</span>
-          <span>{data.post.displayName ?? data.post.user.displayName}</span>
-        </li>
-        <li className='flex flex-wrap gap-1 md:gap-2'>
-          <span className='flex-grow lg:flex-grow-0 font-semibold'>
-            등록일자
-          </span>
-          <span>{DateFormat(data.post.createdAt)}</span>
-        </li>
-        <li className='flex flex-wrap gap-1 md:gap-2'>
-          <span className='flex-grow lg:flex-grow-0 font-semibold'>조회수</span>
-          <span>{data.post.views}</span>
-        </li>
-        <li>
-          <HasLikes postId={postId} />
-        </li>
-      </ul>
+    <Fragment>
+      <DetailHeader postId={postId} data={data} />
 
       <div className='w-full px-5 md:px-0 md:max-w-[1016px] mx-auto pb-[100px]'>
-        {/* image */}
-        <div className='pt-4 w-full'>
-          {data.post.image
-            ? data.post.image?.map((imageURL) => (
-                <div
-                  key={imageURL}
-                  className='relative w-full md:max-w-[800px] mx-auto'
-                >
-                  <AutoHeightImageWrapper
-                    src={imageURL}
-                    alt={`${data.post.user.displayName} 업로드 이미지`}
-                  />
-                </div>
-              ))
-            : null}
-        </div>
+        <DetailImage data={data} />
 
         {/* contents */}
         <div
@@ -158,7 +60,7 @@ const CommDetailPage = ({ postId, data }: CommDetailPageProps) => {
       <Comments postId={postId} />
 
       <PrevNextPost pathname={pathname} prev={data.previous} next={data.next} />
-    </ContainerBox>
+    </Fragment>
   );
 };
