+++ app/components/Posts/PostDetail/DetailHeader/types.ts
@@ -1,0 +1,6 @@
+import { PostData } from '@/apis/posts';
+
+export interface DetailHeaderProps {
+  postId: number;
+  data: PostData;
+}
